+++ deps/v8/src/heap/heap-controller.h
@@ -43,7 +43,7 @@ class V8_EXPORT_PRIVATE MemoryController : public AllStatic {
                               Heap::HeapGrowingMode growing_mode);
 
   static size_t BoundAllocationLimit(Heap* heap, size_t current_size,
-                                     size_t limit, size_t min_size,
+                                     uint64_t limit, size_t min_size,
                                      size_t max_size, size_t new_space_capacity,
                                      Heap::HeapGrowingMode growing_mode);
